requirements.txt
openai>=0.27.0
python-dotenv>=1.1.0
pyPDF2>=3.0.1
langchain>=0.3.25
faiss-cpu>=1.11.0